Prueba

+++ Ulatina.PrograAvanzada.Calentamiento/Examen/Dominio/Acciones/CalcularPerimetro.cs
@@ -7,9 +7,9 @@ namespace Examen.Dominio.Acciones
 {
     public class CalcularPerimetro
     {
-        public int CalcularPerimetroTriangulo(int lado1, int lado2, int lado3)
+        public double CalcularPerimetroTriangulo(double lado1, double lado2, double lado3)
         {
-            int elResultado = 0;
+            double elResultado = 0;
             var laEspecificacion = new Especificaciones.CalculePerimetro();
             elResultado = laEspecificacion.PerimetroTriangulo(lado1, lado2, lado3);
             return elResultado;
